spark agent #169 code review changes
